Extract related logic to a module
